Improve CUDA version checking code (#2448)
* Improve CUDA version checking code

* Use cudaRuntimeGetVersion and cudaDriverGetVersion to check the
  compatibility of the toolkit versions vs driver version
* Prints warning messages if the driver version is not included in the
  CudaToDriverVersion in Debug builds.
* Fix int_version_to_string function

* Fix warnings and format strings in platform code and tests

(cherry picked from commit b89fab7852fa909b99336f1f09ac12ff5d6a1f68)

+++ test/scan.cpp
@@ -272,8 +272,8 @@ TEST(Scan, ExclusiveSum1D) {
     const int in_size = 80000;
     vector<int> h_in(in_size, 1);
     vector<int> h_gold(in_size, 0);
-    for (int i = 1; i < h_gold.size(); ++i) {
-        h_gold[i] = h_in[i] + h_gold[i-1];
+    for (size_t i = 1; i < h_gold.size(); ++i) {
+        h_gold[i] = h_in[i] + h_gold[i - 1];
     }
 
     array in(in_size, &h_in.front());
@@ -286,10 +286,10 @@ TEST(Scan, ExclusiveSum2D_Dim0) {
     const int in_size = 80000 * 2;
     vector<int> h_in(in_size, 1);
     vector<int> h_gold(in_size, 0);
-    for (int i = 1; i < h_gold.size() / 2; ++i) {
+    for (size_t i = 1; i < h_gold.size() / 2; ++i) {
         h_gold[i] = h_in[i] + h_gold[i - 1];
     }
-    for (int i = h_gold.size() / 2 + 1; i < h_gold.size(); ++i) {
+    for (size_t i = h_gold.size() / 2 + 1; i < h_gold.size(); ++i) {
         h_gold[i] = h_in[i] + h_gold[i - 1];
     }
 
@@ -304,10 +304,10 @@ TEST(Scan, ExclusiveSum2D_Dim1) {
     const int in_size = 80000 * 2;
     vector<int> h_in(in_size, 1);
     vector<int> h_gold(in_size, 0);
-    for (int i = 1; i < h_gold.size() / 2; ++i) {
+    for (size_t i = 1; i < h_gold.size() / 2; ++i) {
         h_gold[i] = h_in[i] + h_gold[i - 1];
     }
-    for (int i = h_gold.size() / 2 + 1; i < h_gold.size(); ++i) {
+    for (size_t i = h_gold.size() / 2 + 1; i < h_gold.size(); ++i) {
         h_gold[i] = h_in[i] + h_gold[i - 1];
     }
 
@@ -323,10 +323,10 @@ TEST(Scan, ExclusiveSum2D_Dim2) {
     const int in_size = 80000 * 2;
     vector<int> h_in(in_size, 1);
     vector<int> h_gold(in_size, 0);
-    for (int i = 1; i < h_gold.size() / 2; ++i) {
+    for (size_t i = 1; i < h_gold.size() / 2; ++i) {
         h_gold[i] = h_in[i] + h_gold[i - 1];
     }
-    for (int i = h_gold.size() / 2 + 1; i < h_gold.size(); ++i) {
+    for (size_t i = h_gold.size() / 2 + 1; i < h_gold.size(); ++i) {
         h_gold[i] = h_in[i] + h_gold[i - 1];
     }
 
@@ -339,20 +339,20 @@ TEST(Scan, ExclusiveSum2D_Dim2) {
 }
 
 TEST(Scan, ExclusiveSum2D_Dim3) {
-  const int in_size = 80000 * 2;
-  vector<int> h_in(in_size, 1);
-  vector<int> h_gold(in_size, 0);
-  for (int i = 1; i < h_gold.size() / 2; ++i) {
-    h_gold[i] = h_in[i] + h_gold[i - 1];
-  }
-  for (int i = h_gold.size() / 2 + 1; i < h_gold.size(); ++i) {
-    h_gold[i] = h_in[i] + h_gold[i - 1];
-  }
+    const int in_size = 80000 * 2;
+    vector<int> h_in(in_size, 1);
+    vector<int> h_gold(in_size, 0);
+    for (size_t i = 1; i < h_gold.size() / 2; ++i) {
+        h_gold[i] = h_in[i] + h_gold[i - 1];
+    }
+    for (size_t i = h_gold.size() / 2 + 1; i < h_gold.size(); ++i) {
+        h_gold[i] = h_in[i] + h_gold[i - 1];
+    }
 
-  array in(1, 1, 2, in_size / 2, &h_in.front());
-  array out = scan(in, 3, AF_BINARY_ADD, false);
-  array gold(in_size / 2, 2, &h_gold.front());
-  gold = af::reorder(gold, 2, 3, 1, 0);
+    array in(1, 1, 2, in_size / 2, &h_in.front());
+    array out = scan(in, 3, AF_BINARY_ADD, false);
+    array gold(in_size / 2, 2, &h_gold.front());
+    gold = af::reorder(gold, 2, 3, 1, 0);
 
-  ASSERT_ARRAYS_EQ(gold, out);
+    ASSERT_ARRAYS_EQ(gold, out);
 }
